Add command "ideaproj" which creates a simple IDEA project (and opens it in IDEA if OS X)
It creates a very simple project (and workspace) file which might not work for some setups. It has been tested for IntelliJ IDEA 10.5.1 only. If you are using OS X, the command will also try to open the project in (all!) IDEA apps located in /Applications directory, assuming that you keep only one there. Yes, it is not good, so one needs to adjust the Python script to make it smarter.